+++ src/safir/testing/kubernetes.py
@@ -102,8 +102,13 @@ class _EventStream:
 
     @property
     def next_resource_version(self) -> str:
-        """Resource version of the next event."""
-        return str(len(self._events))
+        """Resource version of the next event.
+
+        This starts with ``1`` to ensure that a resource version of ``0`` is
+        special and means to return all known events, so it must be adjusted
+        when indexing into a list of events.
+        """
+        return str(len(self._events) + 1)
 
     def add_event(self, event: dict[str, Any]) -> None:
         """Add a new event and notify all watchers.
@@ -119,8 +124,10 @@ class _EventStream:
 
     def build_watch_response(
         self,
-        resource_version: str = "0",
+        resource_version: Optional[str] = None,
         timeout_seconds: Optional[int] = None,
+        *,
+        field_selector: Optional[str] = None,
     ) -> Mock:
         """Construct a response to a watch request.
 
@@ -137,14 +144,24 @@ class _EventStream:
             How long to watch. This is total elapsed time, not the time
             between events. It matches the corresponding parameter in the
             Kubernetes API. If `None`, watch forever.
+        field_selector
+            Which events to retrieve when performing a watch. If set, it must
+            be set to ``metadata.name=...`` to match a specific object name.
 
         Returns
         -------
         Mock
             Simulation of an aiohttp response whose ``readline`` method wraps
             the iterator.
+
+        Raises
+        ------
+        AssertionError
+            Some other ``field_selector`` was provided.
         """
-        generator = self._build_watcher(resource_version, timeout_seconds)
+        generator = self._build_watcher(
+            resource_version, timeout_seconds, field_selector
+        )
 
         async def readline() -> bytes:
             return await generator.__anext__()
@@ -156,8 +173,9 @@ class _EventStream:
 
     def _build_watcher(
         self,
-        resource_version: Optional[str] = None,
-        timeout_seconds: Optional[int] = None,
+        resource_version: str | None,
+        timeout_seconds: int | None,
+        field_selector: str | None,
     ) -> AsyncIterator[bytes]:
         """Construct a watcher for this event stream.
 
@@ -171,6 +189,9 @@ class _EventStream:
             How long to watch. This is total elapsed time, not the time
             between events. It matches the corresponding parameter in the
             Kubernetes API. If `None`, watch forever.
+        field_selector
+            Which events to retrieve when performing a watch. If set, it must
+            be set to ``metadata.name=...`` to match a specific object name.
 
         Returns
         -------
@@ -178,10 +199,22 @@ class _EventStream:
             An async iterator that will return each new event in the stream,
             encoded as expected by the Kubernetes API, until the timeout
             occurs.
+
+        Raises
+        ------
+        AssertionError
+            Some other ``field_selector`` was provided.
         """
         timeout = None
         if timeout_seconds is not None:
             timeout = current_datetime() + timedelta(seconds=timeout_seconds)
+
+        # Parse the field selector, if one was provided.
+        name = None
+        if field_selector:
+            match = re.match(r"metadata\.name=(.*)$", field_selector)
+            assert match and match.group(1)
+            name = match.group(1)
 
         # Create and register a new trigger.
         trigger = asyncio.Event()
@@ -189,11 +222,16 @@ class _EventStream:
 
         # Construct the iterator.
         async def next_event() -> AsyncIterator[bytes]:
-            position = int(resource_version or self.next_resource_version)
+            if resource_version:
+                position = int(resource_version)
+            else:
+                position = len(self._events)
             while True:
                 for event in self._events[position:]:
-                    yield json.dumps(event).encode()
                     position += 1
+                    if name and event["object"]["metadata"]["name"] != name:
+                        continue
+                    yield json.dumps(event).encode()
                 if not timeout:
                     await trigger.wait()
                 else:
@@ -272,8 +310,8 @@ class MockKubernetesApi:
         self._nodes = V1NodeList(items=[])
         self._objects: dict[str, dict[str, dict[str, Any]]] = {}
         self._events: defaultdict[str, list[CoreV1Event]] = defaultdict(list)
-        self._event_streams: defaultdict[str, _EventStream]
-        self._event_streams = defaultdict(_EventStream)
+        self._event_streams: defaultdict[str, dict[str, _EventStream]]
+        self._event_streams = defaultdict(lambda: defaultdict(_EventStream))
 
     def get_all_objects_for_test(self, kind: str) -> list[Any]:
         """Return all objects of a given kind sorted by namespace and name.
@@ -645,7 +683,7 @@ class MockKubernetesApi:
         """
         self._maybe_error("create_namespaced_event", namespace, body)
         self._update_metadata(body, "v1", "Event", namespace)
-        stream = self._event_streams[namespace]
+        stream = self._event_streams[namespace]["Event"]
         body.metadata.resource_version = stream.next_resource_version
         stream.add_event({"type": "ADDED", "object": body.to_dict()})
         self._events[namespace].append(body)
@@ -655,7 +693,7 @@ class MockKubernetesApi:
         namespace: str,
         *,
         field_selector: Optional[str] = None,
-        resource_version: str = "0",
+        resource_version: Optional[str] = None,
         timeout_seconds: Optional[int] = None,
         watch: bool = False,
         _preload_content: bool = True,
@@ -701,7 +739,7 @@ class MockKubernetesApi:
         assert not _preload_content
 
         # Return the mock response expected by the Kubernetes API.
-        stream = self._event_streams[namespace]
+        stream = self._event_streams[namespace]["Event"]
         return stream.build_watch_response(resource_version, timeout_seconds)
 
     # NAMESPACE API
@@ -863,7 +901,10 @@ class MockKubernetesApi:
         self._maybe_error("create_namespaced_pod", namespace, body)
         self._update_metadata(body, "v1", "Pod", namespace)
         body.status = V1PodStatus(phase=self.initial_pod_phase)
+        stream = self._event_streams[namespace]["Pod"]
+        body.metadata.resource_version = stream.next_resource_version
         self._store_object(namespace, "Pod", body.metadata.name, body)
+        stream.add_event({"type": "ADDED", "object": body.to_dict()})
         if self.initial_pod_phase == "Running":
             event = CoreV1Event(
                 metadata=V1ObjectMeta(
@@ -899,12 +940,26 @@ class MockKubernetesApi:
             Raised with 404 status if the pod was not found.
         """
         self._maybe_error("delete_namespaced_pod", name, namespace)
-        return self._delete_object(namespace, "Pod", name)
+        pod = self._get_object(namespace, "Pod", name)
+        result = self._delete_object(namespace, "Pod", name)
+        stream = self._event_streams[namespace]["Pod"]
+        stream.add_event({"type": "DELETED", "object": pod.to_dict()})
+        return result
 
     async def list_namespaced_pod(
-        self, namespace: str, *, field_selector: Optional[str] = None
-    ) -> V1PodList:
+        self,
+        namespace: str,
+        *,
+        field_selector: Optional[str] = None,
+        resource_version: Optional[str] = None,
+        timeout_seconds: Optional[int] = None,
+        watch: bool = False,
+        _preload_content: bool = True,
+        _request_timeout: Optional[int] = None,
+    ) -> V1PodList | Mock:
         """List pod objects in a namespace.
+
+        This does support watches.
 
         Parameters
         ----------
@@ -913,36 +968,100 @@ class MockKubernetesApi:
         field_selector
             Only ``metadata.name=...`` is supported. It is parsed to find the
             pod name and only pods matching that name will be returned.
+        resource_version
+            Where to start in the event stream when performing a watch. If
+            `None`, starts with the next change.
+        timeout_seconds
+            How long to return events for before exiting when performing a
+            watch.
+        watch
+            Whether to act as a watch.
+        _preload_content
+            Verified to be `False` when performing a watch.
+        _request_timeout
+            Ignored, accepted for compatibility with the watch API.
 
         Returns
         -------
-        kubernetes_asyncio.client.V1PodList
-            List of pods in that namespace.
+        kubernetes_asyncio.client.V1PodList or unittest.mock.Mock
+            List of pods in that namespace, when not called as a watch. If
+            called as a watch, returns a mock ``aiohttp.Response`` with a
+            ``readline`` metehod that yields the events.
 
         Raises
         ------
-        kubernetes_asyncio.client.ApiException
-            Raised with 404 status if the namespace does not exist.
         AssertionError
             Some other ``field_selector`` was provided.
+        kubernetes_asyncio.client.ApiException
+            Raised with 404 status if the namespace does not exist.
         """
         self._maybe_error("list_namespaced_pod", namespace, field_selector)
         if namespace not in self._objects:
             msg = f"Namespace {namespace} not found"
             raise ApiException(status=404, reason=msg)
-        if field_selector:
-            match = re.match(r"metadata\.name=(.*)$", field_selector)
-            assert match and match.group(1)
-            try:
-                pod = self._get_object(namespace, "Pod", match.group(1))
-                return V1PodList(kind="Pod", items=[pod])
-            except ApiException:
-                return V1PodList(kind="Pod", items=[])
-        else:
-            pods = []
-            for obj in self._objects[namespace]["Pod"].values():
-                pods.append(obj)
-            return V1PodList(kind="Pod", items=pods)
+        if not watch:
+            if field_selector:
+                match = re.match(r"metadata\.name=(.*)$", field_selector)
+                assert match and match.group(1)
+                try:
+                    pod = self._get_object(namespace, "Pod", match.group(1))
+                    return V1PodList(kind="Pod", items=[pod])
+                except ApiException:
+                    return V1PodList(kind="Pod", items=[])
+            else:
+                pods = []
+                for obj in self._objects[namespace]["Pod"].values():
+                    pods.append(obj)
+                return V1PodList(kind="Pod", items=pods)
+
+        # All watches must not preload content since we're returning raw JSON.
+        # This is done by the Kubernetes API Watch object.
+        assert not _preload_content
+
+        # Return the mock response expected by the Kubernetes API.
+        stream = self._event_streams[namespace]["Pod"]
+        return stream.build_watch_response(
+            resource_version, timeout_seconds, field_selector=field_selector
+        )
+
+    async def patch_namespaced_pod_status(
+        self, name: str, namespace: str, body: list[dict[str, Any]]
+    ) -> V1Secret:
+        """Patch the status of a pod object.
+
+        Parameters
+        ----------
+        name
+            Name of pod object.
+        namespace
+            Namespace of secret object.
+        body
+            Patches to apply. Only patches with ``op`` of ``replace`` are
+            supported, and only with ``path`` of ``/status/phase``.
+
+        Returns
+        -------
+        kubernetes_asyncio.client.V1Pod
+            Patched pod object.
+
+        Raises
+        ------
+        kubernetes_asyncio.client.ApiException
+            Raised with 404 status if the secret does not exist.
+        AssertionError
+            Raised if any other type of patch was provided.
+        """
+        self._maybe_error("patch_namespaced_pod", name, namespace)
+        pod = copy.deepcopy(self._get_object(namespace, "Pod", name))
+        for change in body:
+            assert change["op"] == "replace"
+            assert change["path"] == "/status/phase"
+            pod.status.phase = change["value"]
+        stream = self._event_streams[namespace]["Pod"]
+        pod.metadata.resource_version = stream.next_resource_version
+        self._store_object(namespace, "Pod", name, pod, replace=True)
+        stream.add_event({"type": "MODIFIED", "object": pod.to_dict()})
+        return pod
 
     async def read_namespaced_pod(self, name: str, namespace: str) -> V1Pod:
         """Read a pod object.